refactor(common): move macros
